@@ -1,18 +1,23 @@
 module github.com/trntv/sshed
 
-go 1.14
+go 1.23.4
 
 require (
 	github.com/boltdb/bolt v1.3.1
-	github.com/cpuguy83/go-md2man/v2 v2.0.0 // indirect
 	github.com/kevinburke/ssh_config v0.0.0-20180317175531-9fc7bb800b55
-	github.com/mattn/go-colorable v0.1.8 // indirect
 	github.com/mgutz/ansi v0.0.0-20200706080929-d51e80ef957d
-	github.com/pelletier/go-buffruneio v0.2.0 // indirect
 	github.com/pkg/errors v0.8.0
+	github.com/urfave/cli v1.22.5
+	gopkg.in/AlecAivazis/survey.v1 v1.8.8
+)
+
+require (
+	github.com/cpuguy83/go-md2man/v2 v2.0.0 // indirect
+	github.com/kballard/go-shellquote v0.0.0-20180428030007-95032a82bc51 // indirect
+	github.com/mattn/go-colorable v0.1.8 // indirect
+	github.com/mattn/go-isatty v0.0.12 // indirect
+	github.com/pelletier/go-buffruneio v0.2.0 // indirect
 	github.com/russross/blackfriday/v2 v2.1.0 // indirect
 	github.com/stretchr/testify v1.5.1 // indirect
-	github.com/urfave/cli v1.22.5
 	golang.org/x/sys v0.0.0-20210225134936-a50acf3fe073 // indirect
-	gopkg.in/AlecAivazis/survey.v1 v1.8.8
 )
